Use double precision real number to fix oscillator sync

+++ include/Mixer.h
@@ -55,7 +55,7 @@ const int BYTES_PER_SURROUND_FRAME = sizeof( surroundSampleFrame );
 const float OUTPUT_SAMPLE_MULTIPLIER = 32767.0f;
 
 
-const float BaseFreq = 440.0f;
+const double BaseFreq = 440.0;
 const Keys BaseKey = Key_A;
 const Octaves BaseOctave = DefaultOctave;
